WIP: Refactor core: datastore2 -> columnstore; move lots of other things around
