+++ core/src/main/scala/filodb.core/Messages.scala
@@ -1,7 +1,7 @@
-package filodb.core.messages
+package filodb.core
 
 /**
- * Common messages for inter-Actor communication
+ * Common types for asynchronous command/response
  */
 
 // The parent trait for all commands, esp sent down to I/O data storage actors/futures
@@ -14,8 +14,6 @@ trait Response
 trait ErrorResponse extends Response
 
 // Common responses
-case class StorageEngineException(t: Throwable) extends ErrorResponse
-case class MetadataException(t: Throwable) extends ErrorResponse
 case object NoSuchCommand extends ErrorResponse
 case object InconsistentState extends ErrorResponse  // for conditional updates - race condition detected!
 case object TooManyRequests extends ErrorResponse    // Need to retry later when limit dies down
@@ -25,4 +23,7 @@ case object NotApplied extends Response
 case object AlreadyExists extends Response
 case object Success extends Response
 
+// Common exceptions
 case class NotFoundError(what: String) extends Exception
+case class StorageEngineException(t: Throwable) extends Exception
+case class MetadataException(t: Throwable) extends Exception
