+++ core/src/test/scala/filodb.core/columnstore/ChunkRowMapSpec.scala
@@ -1,4 +1,4 @@
-package filodb.core.datastore2
+package filodb.core.columnstore
 
 import java.nio.ByteBuffer
